change data type to unsigned long long

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,26 +8,26 @@
 int main(void)
 {
 	int k;
-	unsigned long int sum;
-	unsigned long int prev;
-	unsigned long int temp;
+	unsigned long long sum;
+	unsigned long long prev;
+	unsigned long long temp;
 
 
 	k = 0;
 	sum = 2;
 	temp = 1;
 	prev = 1;
-	printf("%lu, ", prev);
-	printf("%lu, ", sum);
+	printf("%llu, ", prev);
+	printf("%llu, ", sum);
 	while (k < 96)
 	{
 		temp = sum;
 		sum = sum + prev;
 		prev = temp;
 		if (k != 95)
-			printf("%lu, ", sum);
+			printf("%llu, ", sum);
 		else
-			printf("%lu\n", sum);
+			printf("%llu\n", sum);
 		k++;
 	}
